requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.6.1
sqlmodel==0.0.22
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
bcrypt==4.2.0
httpx==0.27.2
litellm==1.49.2
langchain==0.2.14
langchain-community==0.2.12
chromadb==0.5.11
sentence-transformers==3.0.1
numpy==1.26.4
pypdf==4.3.1
python-docx==1.1.2
docx2txt==0.8
rapidfuzz==3.9.7
tavily-python==0.5.0
python-dotenv==1.0.1

